+++ cmd/redact/git_smudge.go
@@ -1,9 +1,10 @@
 package main
 
 import (
+	"context"
 	"os"
 
-	"github.com/urfave/cli/v2"
+	"github.com/urfave/cli/v3"
 )
 
 func (rt *Runtime) gitSmudgeCmd() *cli.Command {
@@ -16,7 +17,7 @@ func (rt *Runtime) gitSmudgeCmd() *cli.Command {
 	}
 }
 
-func (rt *Runtime) gitSmudgeDo(_ *cli.Context) error {
+func (rt *Runtime) gitSmudgeDo(_ context.Context, _ *cli.Command) error {
 	err := rt.SecretKey.Decode(os.Stdin, os.Stdout)
 	if err != nil {
 		return err