+++ cmd/redact/key_list.go
@@ -1,8 +1,10 @@
 package main
 
 import (
+	"context"
+
 	"github.com/julian7/redact/files"
-	"github.com/urfave/cli/v2"
+	"github.com/urfave/cli/v3"
 )
 
 func (rt *Runtime) keyListCmd() *cli.Command {
@@ -15,7 +17,7 @@ func (rt *Runtime) keyListCmd() *cli.Command {
 	}
 }
 
-func (rt *Runtime) listDo(_ *cli.Context) error {
+func (rt *Runtime) listDo(ctx context.Context, cmd *cli.Command) error {
 	rt.Logger.Infof("repo key: %v", rt.SecretKey)
 
 	return files.EachKey(rt.SecretKey.Keys, func(_ uint32, key files.KeyHandler) error {